+++ examples/current-sensor/src/GUI/WindowIDLE.h
