+++ examples/current-sensor/src/GUI/WindowPack.h
@@ -2,11 +2,23 @@
 #define _WINDOW_PACK_
 
 
-#include "emGUI/emGUI.hpp"
-#include "emGUI/emGUI.h"
-//#include "emGUIGlue.h"
-#include "../../GUI.h"
+//#include "emGUI/emGUI.hpp"
+//#include "emGUI/emGUI.h"
+#include <emGUI.h>
+#include "../../emGUIGlue.h"
 
+
+typedef enum {
+	WINDOW_MENU,
+	WINDOW_MALFUNCTION,
+	WINDOW_TP_TEST,
+	WINDOW_BATTERY,
+	WINDOW_WIFI,
+	WINDOW_IDLE,	//WindowIDLE
+	WINDOW_POPUP,	//WindowPopUp
+	WINDOW_SHORT_MESSAGE,
+	WINDOW_PLOT
+}eWindows;
 
 #define COLOR_BLUE EMGUI_COLOR_SEA_3
 #define so 5
