+++ examples/current-sensor/src/GUI/deviceState.cpp
