+++ examples/current-sensor/src/GUI/deviceState.hpp
